requirements.txt
absl-py==2.0.0

aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bert-for-tf2==0.14.9
bleach==6.0.0
blinker==1.8.2
blis==1.2.0
boto3==1.35.63
botocore
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.5
conllu==6.0.0
contourpy==1.1.1
cryptography==41.0.5
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dnslib==0.9.25
dnspython==2.4.2
# en-core-sci-lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.0/en_core_sci_lg-0.5.0.tar.gz#sha256=351e2d898d1dc73d057f929d2acec137147dff855cedc83261e5eeb6a9188483
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
gast==0.5.4
gensim==4.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.59.3
h11==0.14.0
h5py==3.12.1
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupytext==1.15.2
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
libclang==16.0.6
lxml==4.9.3
marisa-trie==1.2.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
multidict==6.0.4
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
nltk==3.8.1
nmslib-metabrainz==2.1.3
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
openpy==0.0.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
params-flow==0.8.2
parser-libraries==3.7
parso==0.8.3
pathlib_abc==0.1.1
pathy==0.11.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.18.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
py-params==0.10.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.13.6
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.16.1
pymongo==4.5.0
PyMySQL==1.1.0
pyparsing==3.1.1
pysbd==0.3.4
PySocks==1.7.1

pyTelegramBotAPI==4.14.0
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3

PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.9.2
rsa==4.9


scikit-learn==1.3.1


seaborn==0.13.0
selenium==4.14.0
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-nlp==5.5.1
sparknlp==1.0.0
SQLAlchemy==2.0.22
srsly==2.4.8
stack-data==0.6.2
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2

termcolor==2.3.0
terminado==0.17.1
thinc==8.3.4
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizer==3.4.3
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typer==0.4.2
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
wasabi==0.10.1
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
